% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{build_equation}
\alias{build_equation}
\title{Builds the default volume equation as defined by the US
Forest Service configuration datasets.}
\usage{
build_equation(region, spcd)
}
\arguments{
\item{region}{Region code}

\item{spcd}{FIA species code}
}
\value{
An R function that computes CVTS for the input
region and species code. If the code cannot be built,
returns "EE"
}
\description{
Builds the default volume equation as defined by the US
Forest Service configuration datasets.
}
