% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_cvts}
\alias{calc_cvts}
\title{Serves as the default CVTS calculation function,
retrieves the coefficient table for the specified species,
region and volume type, generates the function, and applies
it to the input data.}
\usage{
calc_cvts(dbh, ht, region, spcd)
}
\arguments{
\item{dbh}{A column of diameter at breast height values (or a single value)}

\item{ht}{A column of height values (or a single value)}

\item{region}{Geographic region}

\item{spcd}{FIA species code}
}
\description{
Serves as the default CVTS calculation function,
retrieves the coefficient table for the specified species,
region and volume type, generates the function, and applies
it to the input data.
}
