% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_vol}
\alias{calc_vol}
\title{Serves as the default volume calculation function,
retrieves the coefficient table for the specified species,
region and volume type, generates the function, and applies
it to the input data.}
\usage{
calc_vol(dbh, ht, spcd, region, vol_type)
}
\arguments{
\item{dbh}{A column of diameter at breast height values (or a single value)}

\item{ht}{A column of height values (or a single value)}

\item{spcd}{FIA species code}

\item{region}{Geographic region}
}
\description{
Serves as the default volume calculation function,
retrieves the coefficient table for the specified species,
region and volume type, generates the function, and applies
it to the input data.
}
