% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{common_lookup}
\alias{common_lookup}
\title{Returns the common name of a species
given some input FIA species code}
\usage{
common_lookup(spcd, latin = FALSE)
}
\arguments{
\item{spcd}{The FIA species code}
}
\value{
A string of the species common name
}
\description{
Returns the common name of a species
given some input FIA species code
}
