% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{failing}
\alias{failing}
\title{A convenient way to quickly list the failing species
for a given region}
\usage{
failing(region)
}
\description{
A convenient way to quickly list the failing species
for a given region
}
