% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_CSV}
\alias{find_CSV}
\title{Finds the csv with the input search term
Mostly for use as a debugging tool and internal function}
\usage{
find_CSV(regexp)
}
\arguments{
\item{regexp}{Regex search term as string}
}
\value{
The absolute path to the csv file
}
\description{
Finds the csv with the input search term
Mostly for use as a debugging tool and internal function
}
