% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{get_coefs}
\alias{get_coefs}
\title{Gets the coefficient table for a species in a specific region.}
\usage{
get_coefs(region, spcd)
}
\arguments{
\item{region}{The region code string to retrieve from, for instance
'OR_W' is western Oregon.}

\item{spcd}{The FIA species code}
}
\value{
A 1 row coefficient table containing the values of the coefficients
for the specified region and species.
}
\description{
Gets the coefficient table for a species in a specific region.
}
