% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_config}
\alias{test_config}
\title{Tests for the existence and performance of equation IDs in a given configuration}
\usage{
test_config(config_id)
}
\arguments{
\item{config_id}{The configuration ID specifying the region to test
ex: 'W_OR'}
}
\value{
A dataframe of 'results', either not yet implemented equations or the performance
as calculated by \code{test_eq}
}
\description{
Tests for the existence and performance of equation IDs in a given configuration
}
