% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_config_sp}
\alias{test_config_sp}
\title{Configuration test for species instead of equations,
serves as a function to construct the equations completion
status figure in the readme TODO}
\usage{
test_config_sp(region)
}
\description{
Configuration test for species instead of equations,
serves as a function to construct the equations completion
status figure in the readme TODO
}
