% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_eq}
\alias{test_eq}
\title{Tests a specific CVTS function against the FIA dataset
(test_data) this is meant to be an internal function}
\usage{
test_eq(region, spcd, new_cvts_func)
}
\arguments{
\item{spcd}{The FIA species code}

\item{new_cvts_func}{The new CVTS equation to test as an R expression}

\item{region_string}{The configuration ID specifying the region to test}
}
\value{
The average difference (float) mean(test_cvts - new_cvts)
}
\description{
\code{test_eq} returns the average difference between
the new calculated volumes and the testing dataset.
}
\examples{
test_eq(202, 'W_OR', build_equation('CU202020', get_coefs('W_OR', 202)))
}
